backend/requirements.txt
Django>=3.1.0,<3.2.0
djangorestframework>=3.12.0, <3.13.0
djangorestframework-jwt>=1.11.0, <1.12.0
psycopg2>=2.8.6,<2.9.0
djoser>=2.1.0, <2.2.0
django-filter>=2.3.0, <2.4.0
djangorestframework-jsonapi>=4.0.0, <4.1.0
djangorestframework-simplejwt>=4.5.0, <4.6.0
Pillow>=8.0.0, <8.1.0
drf-yasg>=1.20.0, <1.21.0